Apply PR review suggestions
